requirements.txt
asgiref==3.3.4
ChatterBot==1.0.2
chatterbot-corpus==1.2.0
click==8.0.1
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.3
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.12.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
joblib==1.0.1
mathparse==0.1.2
nltk==3.6.2
packaging==20.9
Pillow==8.2.0
Pint==0.17
psycopg2==2.8.6
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.7.5
pytz==2021.1
PyYAML==3.13
regex==2021.4.4
six==1.16.0
SQLAlchemy==1.2.19
sqlparse==0.4.1
tqdm==4.61.0
whitenoise==5.2.0
